gitlab-ci.yml: Have a doc job
This helps identify if we have actually have a build issue or if we have
a documentation build failure. Notifications sent out are specifying the
name of the build, making things much easier to figure out what actually
failed.

Uses another job w/o the need to run the tests (again).

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>